feat: 의사 정보 제공을 위한 리포지토리 및 dto 작업

+++ src/main/java/com/medicall/domain/member/domain/repository/DiagnosisRepository.java
@@ -1,7 +1,7 @@
 package com.medicall.domain.member.domain.repository;
 
-import com.medicall.domain.member.domain.entity.Diagnosis;
+import com.medicall.domain.medical.domain.entity.Diagnosis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiagnosisRepository extends JpaRepository<Diagnosis, Integer> {
+public interface DiagnosisRepository extends JpaRepository<Diagnosis, Long> {
 }
